producer/requirements.txt
kafka-python
jsonschema
pyyaml
requests
python-dotenv
